feat: Add @lowdefy/plugin-aws as default plugin package.
