+++ .pnp.cjs
@@ -3364,6 +3364,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@lowdefy/operators-nunjucks", "workspace:packages/plugins/operators/operators-nunjucks"],\
             ["@lowdefy/operators-uuid", "workspace:packages/plugins/operators/operators-uuid"],\
             ["@lowdefy/operators-yaml", "workspace:packages/plugins/operators/operators-yaml"],\
+            ["@lowdefy/plugin-aws", "workspace:packages/plugins/plugins/plugin-aws"],\
             ["@lowdefy/plugin-next-auth", "workspace:packages/plugins/plugins/plugin-next-auth"],\
             ["@swc/cli", "virtual:babee6e81435a5d101529cd67f2c6b175f4db37a4ab0b58df15adf73dd11be8917ac14caf44ab4e6882a92c61661055072365b349016e85173e049f006fc2305#npm:0.1.57"],\
             ["@swc/core", "npm:1.2.194"],\
