+++ packages/plugins/blocks/blocks-echarts/src/types.js
@@ -20,8 +20,8 @@ import * as blocks from './blocks.js';
 const icons = {};
 const styles = {};
 Object.keys(blocks).forEach((block) => {
-  icons[block] = blocks[block].meta.icons || [];
-  styles[block] = blocks[block].meta.styles || [];
+  icons[block] = blocks[block].meta.icons ?? [];
+  styles[block] = blocks[block].meta.styles ?? [];
 });
 export default {
   blocks: Object.keys(blocks),
