+++ packages/plugins/plugins/plugin-aws/package.json
@@ -27,8 +27,10 @@
   },
   "type": "module",
   "exports": {
-    ".": "./dist/index.js",
-    "./connections/*": "./dist/connections/*"
+    "./*": "./dist/*",
+    "./blocks": "./dist/blocks.js",
+    "./connections": "./dist/connections.js",
+    "./types": "./dist/types.js"
   },
   "files": [
     "dist/*"
