+++ packages/plugins/plugins/plugin-aws/src/blocks/S3UploadButton/style.less
@@ -15,4 +15,4 @@
 */
 
 @import 'antd/lib/upload/style/index.less';
-@import '../Button/style.less';
+@import 'antd/lib/button/style/index.less';
