+++ packages/plugins/plugins/plugin-aws/src/types.js
@@ -15,11 +15,22 @@
   limitations under the License.
 */
 
+import * as blocks from './blocks.js';
 import * as connections from './connections.js';
 
+const icons = {};
+const styles = {};
+Object.keys(blocks).forEach((block) => {
+  icons[block] = blocks[block].meta.icons ?? [];
+  styles[block] = blocks[block].meta.styles ?? [];
+});
+
 export default {
+  blocks: Object.keys(blocks),
   connections: Object.keys(connections),
+  icons,
   requests: Object.keys(connections)
     .map((connection) => Object.keys(connections[connection].requests))
     .flat(),
+  styles: { default: [], ...styles },
 };
